src/main/java/com/letscode/moveisbattle/controller/RegistrationController.java
package com.letscode.moveisbattle.controller;

import com.letscode.moveisbattle.model.request.RegistrationRequest;
import com.letscode.moveisbattle.service.RegistrationService;
import lombok.AllArgsConstructor;
import org.springframework.web.bind.annotation.*;

@RestController
@RequestMapping(path = "user/registration")
@AllArgsConstructor
public class RegistrationController {

    private final RegistrationService registrationService;

    @PostMapping
    public String register(@RequestBody RegistrationRequest request) {
        return registrationService.register(request);
    }

    @GetMapping(path = "confirm")
    public String confirm(@RequestParam("token") String token) {
        return registrationService.confirmToken(token);
    }

}
